Stricter path validation in Info constructors, doc updates
- Constructors for `FileInformation` and `DirectoryInformation` now throw exceptions if the path does not exist, improving error handling.
- Updated unit tests to expect exceptions for non-existent paths; temporarily disabled MIME type assertion pending a fix.
- Improved XML documentation for clarity and accuracy, including parameter names and exception details.
- Simplified buffer size calculation in `TestFile.cs`.
- Ensured information is loaded after file existence check in `FileObject.cs`.
- Minor parameter renaming and comment corrections for consistency.

+++ Rheo.Storage.Test/Models/TestFile.cs
@@ -46,7 +46,7 @@ namespace Rheo.Storage.Test.Models
             )
         {
             ArgumentNullException.ThrowIfNull(content);
-            var bufferSize = (int)GetBufferSize(Information.Size);
+            var bufferSize = GetBufferSize(Information.Size);
 
             using var memoryStream = new MemoryStream(content);
             using var destStream = new FileStream(
